solving null pointer

+++ custom-E-book/src/main/java/com/example/demo/ebook/controller/book/BookController.java
@@ -55,9 +55,11 @@ public class BookController {
 			String temp_loc = System.getProperty("user.dir")+ "/src/main/resources/static/images/temp";
 			File index = new File(temp_loc);
 			String[]entries = index.list();
-			for(String s: entries){
-			    File currentFile = new File(index.getPath(),s);
-			    currentFile.delete();
+			if(entries!=null) {
+				for(String s: entries){
+				    File currentFile = new File(index.getPath(),s);
+				    currentFile.delete();
+				}
 			}
 			Publisher publisher =(Publisher) session.getAttribute("publisher");
 			List<Book> publisherBooks = service.getPublisherBooks(publisher);
